# docs/api.yml
openapi: "3.0.0"
info:
  title: API for book story system
  version: 0.0.1
servers:
  - url: http://locahost:8080
    description: Optional server description, e.g. Main (production) server
...
paths:
...
  /author/list:
    get:
      description: List all Authors
      operationId: GetAuthorList
      tags:
        - Author
      parameters:
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortingFieldParam'
        - $ref: '#/components/parameters/SortingDirectionParam'
      responses:
        '200':
          $ref: '#/components/responses/GetAuthorList'
  /author:
    post:
      description: Create a author
      operationId: PostAuthor
      tags:
        - Author
      requestBody:
        $ref: '#/components/requestBodies/postAuthor'
      responses:
        '201':
          $ref: '#/components/responses/postAuthor'
  /author/{id}:

    parameters:
      - $ref: '#/components/parameters/IdPathParam'

    get:
      operationId: getAuthorId
      description: get Author by id
      tags:
        - Author
      responses:
        '200':
          $ref: '#/components/responses/getAuthor'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateAuthor
      description: update Author by id
      tags:
        - Author
      requestBody:
        $ref: '#/components/requestBodies/putAuthor'
      responses:
        '200':
          $ref: '#/components/responses/putAuthor'

    delete:
      operationId: deleteAuthorById
      description: delete Author by id
      tags:
        - Author
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /category/list:
    get:
      description: List all Categories
      operationId: GetCategoryList
      tags:
        - Category
      parameters:
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortingFieldParam'
        - $ref: '#/components/parameters/SortingDirectionParam'
      responses:
        '200':
          $ref: '#/components/responses/GetCategoryList'
  /category:
    post:
      description: Create a category
      operationId: PostCategory
      tags:
        - Category
      requestBody:
        $ref: '#/components/requestBodies/postCategory'
      responses:
        '201':
          $ref: '#/components/responses/postCategory'
  /category/{id}:

    parameters:
      - $ref: '#/components/parameters/IdPathParam'

    get:
      operationId: getCategoryId
      description: get Category by id
      tags:
        - Category
      responses:
        '200':
          $ref: '#/components/responses/getCategory'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateCategory
      description: update Category by id
      tags:
        - Category
      requestBody:
        $ref: '#/components/requestBodies/putCategory'
      responses:
        '200':
          $ref: '#/components/responses/putCategory'

    delete:
      operationId: deleteCategoryById
      description: delete Category by id
      tags:
        - Category
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /book/list:
    get:
      description: List all Books
      operationId: GetBookList
      tags:
        - Book
      parameters:
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortingFieldParam'
        - $ref: '#/components/parameters/SortingDirectionParam'
      responses:
        '200':
          $ref: '#/components/responses/GetBookList'
  /book:
    post:
      description: Create a book
      operationId: PostBook
      tags:
        - Book
      requestBody:
        $ref: '#/components/requestBodies/postBook'
      responses:
        '201':
          $ref: '#/components/responses/postBook'
  /book/{id}:

    parameters:
      - $ref: '#/components/parameters/IdPathParam'

    get:
      operationId: getBookId
      description: get Book by id
      tags:
        - Book
      responses:
        '200':
          $ref: '#/components/responses/getBook'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateBook
      description: update Book by id
      tags:
        - Book
      requestBody:
        $ref: '#/components/requestBodies/putBook'
      responses:
        '200':
          $ref: '#/components/responses/putBook'

    delete:
      operationId: deleteBookById
      description: delete Book by id
      tags:
        - Book
      responses:
        '200':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    PageDto:
      description: List of PublisherDto with pagination info
      required:
        - page
        - totalPages
        - pageSize
      properties:
        page:
          description: Current page
          type: integer
          format: int32
        totalPages:
          description: Total page
          type: integer
          format: int32
        pageSize:
          description: Quantity of records per page
          type: integer
          format: int32
        totalRows:
          description: total numbers of rows for the request
          type: integer
          format: int64
        sortField:
          description: sorting field
          type: string
        sortDir:
          description: sorting direction
          type: string
          enum: [ASC, DESC]
    PublisherDto:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AuthorDto:
      type: object
      required:
        - firstName
        - secondName
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        secondName:
          type: string
    CategoryDto:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ConditionDto:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    BookDto:
      type: object
      required:
        - title
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        publicationYear:
          type: integer
          format: int32
        price:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/AuthorDto'
        publisher:
          $ref: '#/components/schemas/PublisherDto'
        condition:
          $ref: '#/components/schemas/ConditionDto'
        category:
          $ref: '#/components/schemas/CategoryDto'
...
  requestBodies:
...
    postAuthor:
      description: Body for creating Author
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDto'
    putAuthor:
      description: Body for updating Author
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDto'
...
    postCategory:
      description: Body for creating Category
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDto'
    putCategory:
      description: Body for updating Category
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDto'
...
    postBook:
      description: Body for creating Book
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDto'
    putBook:
      description: Body for updating Book
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDto'
...
  responses:
...
    getAuthor:
      description: Get dto of Author record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDto'
    postAuthor:
      description: Author successfully created. Return dto of created record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDto'
    putAuthor:
      description: Body for updating Author
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDto'
    GetAuthorList:
      description: List of AuthorDto with pagination info
      content:
        application/json:
          schema:
            type: object
            properties:
              records:
                description: List of records to show
                type: array
                items:
                  $ref: '#/components/schemas/AuthorDto'
            allOf:
              - $ref: '#/components/schemas/PageDto'
    getCategory:
      description: Get dto of Category record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDto'
    postCategory:
      description: Category successfully created. Return dto of created record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDto'
    putCategory:
      description: Body for updating Category
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CategoryDto'
    GetCategoryList:
      description: List of AuthorDto with pagination info
      content:
        application/json:
          schema:
            type: object
            properties:
              records:
                description: List of records to show
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDto'
            allOf:
              - $ref: '#/components/schemas/PageDto'
...
    getBook:
      description: Get dto of Book record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDto'
    postBook:
      description: Book successfully created. Return dto of created record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDto'
    putBook:
      description: Body for updating Book
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDto'
    GetBookList:
      description: List of ConditionDto with pagination info
      content:
        application/json:
          schema:
            type: object
            properties:
              records:
                description: List of records to show
                type: array
                items:
                  $ref: '#/components/schemas/BookDto'
            allOf:
              - $ref: '#/components/schemas/PageDto'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                description: Message what exactly not found
                type: string
              content:
                description: Optional content
                type: object
  parameters:
    PageNumberParam:
      in: query
      description: Page number
      name: pageNumber
      required: false
      schema:
        type: integer
        format: int32
        default: 0
    PageSizeParam:
      in: query
      description: Page size
      name: pageSize
      required: false
      schema:
        type: integer
        format: int32
        default: 10
    SortingFieldParam:
      in: query
      description: Sorting field
      name: sortBy
      required: false
      schema:
        type: string
        default: id
    SortingDirectionParam:
      in: query
      description: Sorting direction
      name: sortDir
      required: false
      schema:
        type: string
        enum: ["ASC", "DESC"]
        default: "ASC"
    IdPathParam:
      in: path
      description: Id of record to operate
      name: id
      required: true
      schema:
        type: integer
        format: int64